+++ spines/bullet_spine.cpp
@@ -12,41 +12,27 @@
 #include <string>
 #include <vector>
 
+#include "spines/common/observers.h"
+#include "spines/common/sensors.h"
 #include "upkie/cpp/actuation/BulletInterface.h"
 #include "upkie/cpp/controllers/ControllerPipeline.h"
-#include "upkie/cpp/observers/BaseOrientation.h"
-#include "upkie/cpp/observers/FloorContact.h"
 #include "upkie/cpp/observers/ObserverPipeline.h"
-#include "upkie/cpp/observers/WheelOdometry.h"
-#include "upkie/cpp/sensors/CpuTemperature.h"
 #include "upkie/cpp/sensors/SensorPipeline.h"
 #include "upkie/cpp/spine/Spine.h"
 #include "upkie/cpp/utils/clear_shared_memory.h"
 #include "upkie/cpp/utils/get_log_path.h"
 #include "upkie/cpp/version.h"
 
-#ifndef __APPLE__
-#include "upkie/cpp/sensors/Joystick.h"
-#endif
-
 namespace spines::bullet {
 
 using palimpsest::Dictionary;
 using upkie::cpp::actuation::BulletInterface;
 using upkie::cpp::controllers::ControllerPipeline;
-using upkie::cpp::observers::BaseOrientation;
-using upkie::cpp::observers::FloorContact;
 using upkie::cpp::observers::ObserverPipeline;
-using upkie::cpp::observers::WheelOdometry;
-using upkie::cpp::sensors::CpuTemperature;
 using upkie::cpp::sensors::SensorPipeline;
 using upkie::cpp::spine::Spine;
 using upkie::cpp::utils::clear_shared_memory;
 using upkie::cpp::utils::get_log_path;
-
-#ifndef __APPLE__
-using upkie::cpp::sensors::Joystick;
-#endif
 
 //! Command-line arguments for the Bullet spine.
 class CommandLineArguments {
@@ -172,52 +158,6 @@ class CommandLineArguments {
   bool version = false;
 };
 
-/*! Connect sensors to the observation pipeline.
- *
- * \param[in] interface Actuation interface.
- */
-SensorPipeline make_sensors(const BulletInterface& interface) {
-  SensorPipeline sensors;
-
-  auto cpu_temperature = std::make_shared<CpuTemperature>();
-  sensors.connect_sensor(cpu_temperature);
-
-#ifndef __APPLE__
-  auto joystick = std::make_shared<Joystick>();
-  if (joystick->present()) {
-    spdlog::info("Joystick found");
-    sensors.connect_sensor(joystick);
-  }
-#endif
-
-  return sensors;
-}
-
-/*! Append observers (in a given order) to the observation pipeline.
- *
- * \param[in] spine_frequency Spine frequency in [Hz].
- */
-ObserverPipeline make_observers(unsigned spine_frequency) {
-  ObserverPipeline observation;
-
-  BaseOrientation::Parameters base_orientation_params;
-  auto base_orientation =
-      std::make_shared<BaseOrientation>(base_orientation_params);
-  observation.append_observer(base_orientation);
-
-  FloorContact::Parameters floor_contact_params;
-  floor_contact_params.dt = 1.0 / spine_frequency;
-  auto floor_contact = std::make_shared<FloorContact>(floor_contact_params);
-  observation.append_observer(floor_contact);
-
-  WheelOdometry::Parameters odometry_params;
-  odometry_params.dt = 1.0 / spine_frequency;
-  auto odometry = std::make_shared<WheelOdometry>(odometry_params);
-  observation.append_observer(odometry);
-
-  return observation;
-}
-
 /*! Create the actuation interface.
  *
  * \param[in] argv0 Name of spine binary from the command line.
@@ -265,8 +205,9 @@ int run_spine(const char* argv0, const CommandLineArguments& args) {
   spdlog::info("Spine data logged to {}", params.log_path);
 
   BulletInterface interface = make_actuation_interface(argv0, args);
-  SensorPipeline sensors = make_sensors(interface);
-  ObserverPipeline observers = make_observers(args.spine_frequency);
+  SensorPipeline sensors =
+      common::make_sensors(/* joystick_required = */ false);
+  ObserverPipeline observers = common::make_observers(args.spine_frequency);
   ControllerPipeline controllers;
   Spine spine(params, interface, sensors, observers, controllers);
   if (args.nb_substeps == 0u) {