+++ spines/mock_spine.cpp
@@ -13,11 +13,7 @@
 
 #include "upkie/cpp/actuation/MockInterface.h"
 #include "upkie/cpp/controllers/ControllerPipeline.h"
-#include "upkie/cpp/observers/BaseOrientation.h"
-#include "upkie/cpp/observers/FloorContact.h"
 #include "upkie/cpp/observers/ObserverPipeline.h"
-#include "upkie/cpp/observers/WheelOdometry.h"
-#include "upkie/cpp/sensors/CpuTemperature.h"
 #include "upkie/cpp/sensors/SensorPipeline.h"
 #include "upkie/cpp/spine/Spine.h"
 #include "upkie/cpp/utils/get_log_path.h"
@@ -28,16 +24,14 @@
 #include "upkie/cpp/sensors/Joystick.h"
 #endif
 
+#include "spines/common/observers.h"
+
 namespace spines::mock {
 
 using palimpsest::Dictionary;
 using upkie::cpp::actuation::MockInterface;
 using upkie::cpp::controllers::ControllerPipeline;
-using upkie::cpp::observers::BaseOrientation;
-using upkie::cpp::observers::FloorContact;
 using upkie::cpp::observers::ObserverPipeline;
-using upkie::cpp::observers::WheelOdometry;
-using upkie::cpp::sensors::CpuTemperature;
 using upkie::cpp::sensors::SensorPipeline;
 using upkie::cpp::spine::Spine;
 
@@ -125,40 +119,9 @@ int run_spine(const CommandLineArguments& args) {
     return -4;
   }
 
-  SensorPipeline sensors;
-
-  // Sensor: CPU temperature
-  auto cpu_temperature = std::make_shared<CpuTemperature>();
-  sensors.connect_sensor(cpu_temperature);
-
-#ifndef __APPLE__
-  // Sensor: Joystick
-  auto joystick = std::make_shared<Joystick>();
-  if (joystick->present()) {
-    spdlog::info("Joystick found");
-    sensors.connect_sensor(joystick);
-  }
-#endif
-
-  ObserverPipeline observers;
-
-  // Observation: Base orientation
-  BaseOrientation::Parameters base_orientation_params;
-  auto base_orientation =
-      std::make_shared<BaseOrientation>(base_orientation_params);
-  observers.append_observer(base_orientation);
-
-  // Observation: Floor contact
-  FloorContact::Parameters floor_contact_params;
-  floor_contact_params.dt = 1.0 / args.spine_frequency;
-  auto floor_contact = std::make_shared<FloorContact>(floor_contact_params);
-  observers.append_observer(floor_contact);
-
-  // Observation: Wheel odometry
-  WheelOdometry::Parameters odometry_params;
-  odometry_params.dt = 1.0 / args.spine_frequency;
-  auto odometry = std::make_shared<WheelOdometry>(odometry_params);
-  observers.append_observer(odometry);
+  SensorPipeline sensors =
+      common::make_sensors(/* joystick_required = */ false);
+  ObserverPipeline observers = common::make_observers(args.spine_frequency);
 
   // Mock actuators
   const double dt = 1.0 / args.spine_frequency;